# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresa",
        "description": "Documentação da API de Empresas, Funcionários e Clientes",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/funcionarios": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Lista todos os funcionários",
                "description": "Retorna uma lista paginada de funcionários",
                "operationId": "getFuncionariosList",
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Cadastra um novo funcionário",
                "description": "Cria um novo registro de funcionário",
                "operationId": "storeFuncionario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Funcionário"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "func@example.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "111.222.333-44"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "empresa_ids[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Funcionário criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/funcionarios/{id}": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Detalhes de um funcionário",
                "description": "Retorna os dados de um funcionário específico",
                "operationId": "getFuncionarioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Atualiza um funcionário",
                "description": "Atualiza os dados de um funcionário existente (Use POST com _method=PUT para upload)",
                "operationId": "updateFuncionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "João Funcionário Alterado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "func@example.com"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "111.222.333-44"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "empresa_ids[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funcionário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Remove um funcionário",
                "description": "Deleta o registro de um funcionário do sistema",
                "operationId": "deleteFuncionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Funcionário removido com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": [
            []
        ]
    }
}